fix: type into focused iOS inputs without AX
